partial enemy chasing behavior
The enemy currently will follow a normal walking path, and then go out of it's way to hit the player in the event that the player enter's the slime's detection area. HOWEVER, it will enter a warped path of it's original idle walk state instead of continuous pursuance of the player. Changes to be made